% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMAQmap.R
\name{CMAQ_map}
\alias{CMAQ_map}
\title{summarize county level MOVESoutput}
\usage{
CMAQ_map(ACON, path1, GRID, path2, poll, TS, LAY, outpath)
}
\arguments{
\item{path1:}{path to the ACON file}

\item{ACON:}{target netcdf file}

\item{path2:}{path to the GRID file}

\item{GRID:}{GRIDCRO2D file}

\item{GRID:}{GRIDCRO2D file}

\item{poll:}{pollytant or variable}

\item{TS:}{Timestemp}

\item{LAY:}{Layer}

\item{outpath:}{outpath}
}
\description{
This function summarize county level MOVESouput
}
\examples{


}
