% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stat.R
\name{data_stat}
\alias{data_stat}
\title{summarize county level MOVESoutput}
\usage{
data_stat(file, path, reportfile)
}
\arguments{
\item{path:}{path to the file}

\item{file:}{target netcdf file}

\item{reportfile:}{file for stat report}
}
\description{
This function summarize county level MOVESouput
}
\examples{

}
