% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BC_VIR.R, R/IC_BC_VIR_v01.R
\name{IC_BC_VIR}
\alias{IC_BC_VIR}
\title{summarize county level MOVESoutput}
\usage{
IC_BC_VIR(path, BC, IC, BDY, poll, lay, outpath, country, xlimit)

IC_BC_VIR(path, BC, IC, BDY, poll, lay, outpath, country, xlimit)
}
\arguments{
\item{path:}{path to the file}

\item{BC:}{boundary condition netcdf file}

\item{IC:}{initial condition netcdf file}

\item{poll:}{pollution/species}

\item{lay:}{layer}

\item{outfile:}{outfile name}

\item{path:}{path to the file}

\item{BC:}{boundary condition netcdf file}

\item{IC:}{initial condition netcdf file}

\item{BDY:}{boundary condition GRID file}

\item{poll:}{pollution/species}

\item{lay:}{layer}

\item{outfile:}{outfile name}

\item{country:}{domain region}

\item{xlimit:}{range of longitude}
}
\description{
This function summarize county level MOVESouput

This function summarize county level MOVESouput
}
\examples{




}
