% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{Mask}
\alias{Mask}
\title{summarize county level MOVESoutput}
\usage{
Mask(path, NC1, GRID, outfile, REAL, poll, t, mask_name)
}
\arguments{
\item{path:}{path to the file}

\item{NC1:}{input netcdf file}

\item{GRID:}{GRID file}

\item{outfile:}{outfile with mask point}

\item{REAL:}{the real mask point in csv file}

\item{poll:}{pollution/species}

\item{t:}{time}

\item{mask_name:}{mask variable name}
}
\description{
This function summarize county level MOVESouput
}
\examples{


}
