% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshift.R
\name{tshift}
\alias{tshift}
\title{summarize county level MOVESoutput}
\usage{
tshift(file, path, source_time, target_time, outfile)
}
\arguments{
\item{path:}{path to the file}

\item{file:}{target netcdf file}

\item{var:}{the variable you would like to change}

\item{ratio:}{the ratio you would like to apply}
}
\description{
This function summarize county level MOVESouput
}
\examples{

}
