% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_add.R
\name{var_add}
\alias{var_add}
\title{summarize county level MOVESoutput}
\usage{
var_add(file, path, inpoll, outpoll, outfile, ratio)
}
\arguments{
\item{path:}{path to the file}

\item{file:}{target netcdf file}

\item{var:}{the variable you would like to change}

\item{ratio:}{the ratio you would like to apply}
}
\description{
This function summarize county level MOVESouput
}
\examples{


}
