% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_CMAQdata.R
\name{var_change}
\alias{var_change}
\title{summarize county level MOVESoutput}
\usage{
var_change(file, path, var, ratio)
}
\arguments{
\item{path:}{path to the file}

\item{file:}{target netcdf file}

\item{var:}{the variable you would like to change}

\item{ratio:}{the ratio you would like to apply}
}
\description{
This function summarize county level MOVESouput
}
\examples{


}
