% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_BC_VIR.R
\name{IC_BC_VIR}
\alias{IC_BC_VIR}
\title{summarize county level MOVESoutput}
\usage{
IC_BC_VIR(path, BC, IC, poll, lay, outpath)
}
\arguments{
\item{path:}{path to the file}

\item{BC:}{boundary condition netcdf file}

\item{IC:}{initial condition netcdf file}

\item{poll:}{pollution/species}

\item{lay:}{layer}

\item{outfile:}{outfile name}
}
\description{
This function summarize county level MOVESouput
}
\examples{


}
